avformat: Add Dynacolor MVC Demuxer
This demuxer adds support for demuxing files in the Dynacolor format
such as the sample located at:

http://samples.ffmpeg.org/camera-dvr/dynacolor/dynacolor-camera-sample

However some decode errors are showing on the resulting MPEG4 stream.
I don't know whether this is a bug with the demuxer or the file as there is only one sample
but the output results in a 1 second mp4 file that is playable in VLC media player.

Signed-off-by: Tom Needham <06needhamt@gmail.com>